requirements.txt
streamlit
pandas
numpy
scipy
sqlalchemy
psycopg2-binary
requests
python-dotenv
